+++ docs/javascripts/mathjax.js
@@ -1,0 +1,14 @@
+window.MathJax = {
+  tex: {
+    inlineMath: [["\\(", "\\)"]],
+    displayMath: [["\\[", "\\]"]],
+  },
+  options: {
+    ignoreHtmlClass: "[a-zA-Z]+_html",
+    processHtmlClass: "arithmatex",
+  },
+};
+
+document$.subscribe(() => {
+  MathJax.typesetPromise();
+});